+++ packages/gasket-intl/lib/index.js
@@ -1,12 +1,9 @@
 import { IntlManager } from './intl-manager.js';
 import { LocaleFileStatus } from './constants.js';
 
-/** @type {import('./types').makeIntlManager } */
+/** @type {import('@gasket/intl').makeIntlManager } */
 function makeIntlManager(manifest) {
   return new IntlManager(manifest);
 }
 
-export {
-  makeIntlManager,
-  LocaleFileStatus
-};
+export { makeIntlManager, LocaleFileStatus };
