+++ packages/gasket-intl/package.json
@@ -67,7 +67,10 @@
         ],
         "extends": [
           "godaddy-typescript"
-        ]
+        ],
+        "rules": {
+          "no-use-before-define": "warn"
+        }
       }
     ]
   }
